marks update bug fix

+++ src/main/java/org/ideoholic/curium/model/marksdetails/service/MarksDetailsService.java
@@ -508,7 +508,7 @@ public class MarksDetailsService {
 					marks.setExamid(examid);
 					marks.setSubid(subid);
 					marks.setSid(studentId);
-					marks.setMarksobtained(Integer.parseInt(marksObtained));
+					marks.setMarksobtained(Float.parseFloat(marksObtained));
 					String currentAcademicYear = strCurrentAcademicYear;
 					String currentYear = currentAcademicYear;
 					marks.setAcademicyear(currentYear);
